requirements.txt
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
outcome==1.3.0.post0
pillow==11.1.0
PySocks==1.7.1
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
